requirements/base_reqs.txt
Django==1.6.4
Markdown==2.4
South==0.8.4
argparse==1.2.1
cffi==0.8.2
cryptography==0.4
distribute==0.6.24
django-debug-toolbar==1.0.1
django-devserver==0.8.0
django-filter==0.7
django-model-utils==2.0.3
django-extensions==1.3.7
django-pipeline==1.3.24
djangorestframework==2.3.13
futures==2.1.6
ipdb==0.8
ipython==1.0.0
psycopg2==2.5.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
python-dateutil==2.2
python-magic==0.4.6
pytz==2014.3
service-identity==0.2
sqlparse==0.1.11
wsgiref==0.1.2
zope.interface==4.1.1
